`tv-script-generation`, Remove dropout layer because we don’t need to worry about overfitting the model for this particular project. The dropout layer will just make it harder to train.
- also reduce `rnn_size` hyperparameter to 256 from 300. We want the
`rss_size` to be lower than the `embed_size`

rnn_size should not be more than embed_dim. Because we are using word
embeddings to reduce the input word dimensionality. If we increase the
units in the next layer (which is also input representation) by large
amount then it defeats the purpose of reducing the dimensionality by
the embedding projection layer.

Refer to Figure 1: A neural language model (Bengio et al., 2006)
http://blog.aylien.com/overview-word-embeddings-history-word2vec-cbow-gl
ove/

Embedding Layer: This layer generates word embeddings by multiplying an
index vector with a word embedding matrix;
Intermediate Layer(s): One or more layers that produce an intermediate
representation of the input, e.g. a fully-connected layer that applies
a non-linearity to the concatenation of word embeddings of n previous
words;
Softmax Layer: The final layer that produces a probability distribution
over words in V. Bengio et al. also determine two issues with current
state-of-the-art-models:
– The first is that Layer 2. is replaceable with an LSTM, which is used
by state-of-the-art neural language models
If you think about it in CNN, if we are stacking more conv layers then
number of parameters in each layer decreases. Same should be achieved
by embed_dim and the following layer's units.